% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpSPA.R
\name{EmpSPA.one.SNP.hete}
\alias{EmpSPA.one.SNP.hete}
\title{Empirical SaddlePoint Approximation implementation of logistic regression case-control analysis, Cox regression surival analysis or analysis of other traits (One-SNP-version for heterogeneous case)}
\usage{
EmpSPA.one.SNP.hete(
  g,
  obj.null,
  impute.method = "fixed",
  missing.cutoff = 0.15,
  min.maf = 1e-04,
  G.model = "Add"
)
}
\arguments{
\item{g}{a numeric genotype vector. Missing genotype should be coded as NA. Both hard-called and imputed genotype data are supported.}

\item{others}{the same as function EmpSPA.homo. NOTE that we do not check subject order in this one-snp-version !!!}
}
\value{
the same as function EmpSPA.nonhomo.
}
\description{
One-SNP-version EmpSPA function. This function is to facilitate users that prefer reading and analyzing genotype line-by-line.
}
