% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpSPA.R
\name{EmpSPA_Null_Model}
\alias{EmpSPA_Null_Model}
\title{Fits a NULL model for EmpSPA}
\usage{
EmpSPA_Null_Model(
  traits = "survival/binary/others",
  formula = NULL,
  data = NULL,
  pIDs = NULL,
  gIDs = NULL,
  range = c(-100, 100),
  length.out = 10000,
  Cova.mtx = NULL,
  resid = NULL,
  ...
)
}
\arguments{
\item{traits}{a character value corresponding to phenotype. It should be "binary" for binary phenotype, "survival" for time-to-event phenotype, and "others" for other types of phenotype.}

\item{formula}{a formula to be passed to function glm() or coxph(). For more details, please refer to package survival.}

\item{data}{a data.frame in which to interpret the variables named in the formula}

\item{pIDs}{a character vector of subject IDs. NOTE: its order should be the same as the subjects order in the formula.}

\item{gIDs}{a character vector of subject IDs. NOTE: its order should be the same as the subjects order of the Geno.mtx (i.e. the input of the function EmpSPA()).}

\item{range}{a two-element numeric vector (default: c(-100,100)) to specify the domain of the empirical CGF.}

\item{length.out}{a positive integer (default: 9999) for empirical CGF. Larger length.out corresponds to longer calculation time and more accurate estimated empirical CGF.}

\item{Cova.mtx}{a covariate matrix including a column of 1s (only needed when the type of phenotype is neither survival nor binary).}

\item{resid}{a vector of residuals from a generalized linear regression model in which Cova.mtx is the covariate matrix (only needed when the type of phenotype is neither survival nor binary).}

\item{...}{Other arguments passed to function glm() or coxph(). For more details, please refer to package survival.}
}
\value{
an object with a class of "binary_Null_Model" for binary phenotype or "survival_Null_Model" for time-to-event phenotype.
}
\description{
Fits a null logistic regression model for binary phenotype or a null Cox proportional hazards model for time-to-event phenotype and then calculates the raw residuals or martingale residuals and the empirical cumulant generation function (CGF) of the raw residuals or martingale residuals,
or use residuals from a null generalized linear model to calculate its the empirical CGF.
}
\examples{
Please check help(EmpSPA) for a simulated example.
}
