% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpSPA.R
\name{GRM_PCA}
\alias{GRM_PCA}
\title{Calculate principle components from an empirical genomic relationship matrix (GRM) when all samples are independent.}
\usage{
GRM_PCA(Geno.mtx)
}
\arguments{
\item{Geno.mtx}{a numeric genotype matrix with each row as an individual and each column as a genetic variant.}
}
\value{
top 10 PCs
}
\description{
Use an empirical genomic relationship matrix (GRM) to calculate principle components when all samples are independent.
}
\examples{

# example 1  time-to-event phenotype (non-homogeneous)
# Simulation phenotype and genotype
N = 4000
N1 = N/2
N2 = N/2
nSNP = 10000
p.PCA = runif(nSNP, min = 0.2, max = 0.5)
p1.PCA = runif(nSNP, min = 0.2, max = 0.5)
p2.PCA = runif(nSNP, min = 0.2, max = 0.5)
P.PCA = cbind(p1.PCA, p2.PCA)
### i.i.d subject & i.i.d SNPS matrix
P1.mtx.PCA = matrix(0, N1, nSNP)
for (i in 1:N1) {
 P1.mtx.PCA[i,]= p1.PCA
}
P2.mtx.PCA = matrix(0, N2, nSNP)
for (i in 1:N2) {
 P2.mtx.PCA[i,]= p2.PCA
}
P.mtx.PCA = rbind(P1.mtx.PCA, P2.mtx.PCA) ### true MAF

Geno.mtx1.PCA = matrix(0, N1, nSNP)
for (i in 1:N1) {
 Geno.mtx1.PCA[i,]= rbinom(nSNP, 2, p1.PCA)
}
Geno.mtx2.PCA = matrix(0, N2, nSNP)
for (i in 1:N2) {
 Geno.mtx2.PCA[i,]= rbinom(nSNP, 2, p2.PCA)
}
Geno.mtx.PCA = rbind(Geno.mtx1.PCA, Geno.mtx2.PCA)

GRM_PCA(Geno.mtx.PCA)

}
